authenctication and transform api

+++ src/controller/auth/user.delete.controller.ts
@@ -1,11 +1,17 @@
+import employeeModel from "../../model/employee.model";
 import UserModel from "../../model/user.model"
 import { Request, Response } from "express";
 
 export const deleteUser = async (req:Request, res:Response) => {
   const { id } = req.params;
   try {
-    const user = await UserModel.findOne({ _id: id});
+    const user = await UserModel.findOne({ employeeId: id});
+    const employee = await employeeModel.findOne({ _id: id})
+
     if(!user) throw Error("user not found")
+    if(!employee) throw Error("employee not found")
+    employee.role = "NOT_USER"
+    employee.save()
     await UserModel.findByIdAndDelete(user._id)
     res
       .status(200)
